Refactoring the project structure in the data layer
