+++ app/src/main/java/com/gabrielribeiro/desafio_mobile/repositories/MovieRepository.kt
@@ -1,7 +1,7 @@
 package com.gabrielribeiro.desafio_mobile.repositories
 
 import androidx.lifecycle.LiveData
-import com.gabrielribeiro.desafio_mobile.data.entity.MovieEntity
+import com.gabrielribeiro.desafio_mobile.data.local.entity.MovieEntity
 import com.gabrielribeiro.desafio_mobile.data.remote.models.MoviesListResponse
 import com.gabrielribeiro.desafio_mobile.utils.Resource
