+++ app/src/main/java/com/gabrielribeiro/desafio_mobile/ui/viewmodels/MovieParentViewModel.kt
@@ -2,7 +2,7 @@ package com.gabrielribeiro.desafio_mobile.ui.viewmodels
 
 import android.util.Log
 import androidx.lifecycle.*
-import com.gabrielribeiro.desafio_mobile.data.entity.MovieEntity
+import com.gabrielribeiro.desafio_mobile.data.local.entity.MovieEntity
 import com.gabrielribeiro.desafio_mobile.data.remote.models.MoviesListResponse
 import com.gabrielribeiro.desafio_mobile.repositories.MovieRepository
 import com.gabrielribeiro.desafio_mobile.utils.Resource
